Refactor authentication context: move to lib/database, update imports, and streamline user state initialization
