+++ src/components/common/protected-routes.tsx
@@ -1,9 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { useAuth } from "./auth-context";
-
-// type ProtectedRoutesProps = {
-//   children: React.ReactNode;
-// };
+import { useAuth } from "../../lib/database/auth-context";
 
 export const ProtectedRoutes = ({
   children,
@@ -11,6 +7,5 @@ export const ProtectedRoutes = ({
   children: React.ReactNode;
 }) => {
   const { user } = useAuth();
-  // const isAuthenticated = localStorage.getItem("authUser");
   return user ? <>{children}</> : <Navigate to="/login" />;
 };
