+++ src/components/team/Udoy/sign-up.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from "@/components/common/auth-context";
+import { useAuth } from "@/lib/database/auth-context";
 import { cn } from "@/lib/utils/cn";
 import { FormEvent, useState, type FC } from "react";
 import { useNavigate } from "react-router-dom";
