+++ src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import { App } from "./App.tsx";
 import "./index.css";
-import { AuthProvider } from "./components/common/auth-context.tsx";
+import { AuthProvider } from "./lib/database/auth-context.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
